src/lib/requirements.txt
supabase
python-dotenv
requests
httpx
fastapi==0.110.0
uvicorn==0.27.1
docker==7.0.0 